+++ src/alappviewer/alAppFileViewer.ts
@@ -8,45 +8,82 @@ import { PageBuilder } from '../objectbuilders/pageBuilder';
 import { ObjectBuildersCollection } from '../objectbuilders/objectBuildersCollection';
 import { ALObjectRunner } from '../alObjectRunner';
 
-export class ALAppFileContentProvider implements vscode.TextDocumentContentProvider {
-    onDidChange?: vscode.Event<vscode.Uri>;
-
+export class ALAppFileViewer {
     private objectLibraries : ALObjectLibrariesCollection;
     private mainTemplate : string;
     private codeOutlineProvider : ALOutlineProvider;
     private objectBuilders : ObjectBuildersCollection;
     private alObjectRunner : ALObjectRunner;
+    private htmlContent : string;
+    private panel : vscode.WebviewPanel | undefined;
+    private title : string;
+    private disposables : vscode.Disposable[] = [];
 
-    constructor(context : vscode.ExtensionContext, objLibraries : ALObjectLibrariesCollection, objBuilders : ObjectBuildersCollection) {
+    constructor(context : vscode.ExtensionContext, objLibraries : ALObjectLibrariesCollection, objBuilders : ObjectBuildersCollection,
+        outlineProvider : ALOutlineProvider, objectRunner : ALObjectRunner, uri : vscode.Uri) {
         this.objectLibraries = objLibraries;
         this.objectBuilders = objBuilders;
+        this.codeOutlineProvider = outlineProvider;
+        this.alObjectRunner = objectRunner;
         this.initTemplate(context);
+        this.htmlContent = this.getHtmlContent(uri);
+        this.title = path.parse(uri.fsPath).base;
+        this.panel = undefined;        
     }
 
-    setOutlineProvider(newValue : ALOutlineProvider) {
-        this.codeOutlineProvider = newValue;
-    }
+    dispose() {
+        if (this.panel)
+            this.panel.dispose();
 
-    setALObjectRunner(newValue : ALObjectRunner) {
-        this.alObjectRunner = newValue;
+        while (this.disposables.length) {
+            const x = this.disposables.pop();
+            if (x)
+                x.dispose();
+        }
     }
 
     private initTemplate(context: vscode.ExtensionContext) {
+        var extensionPath = vscode.Uri.file(context.extensionPath).with({scheme : 'vscode-resource' });
         var fs = require('fs');
         var filePath = context.asAbsolutePath(path.join('htmlresources', 'objectbrowser', 'objectbrowser.html'));
         var content = fs.readFileSync(filePath, 'utf8');
-        this.mainTemplate = content.replace(new RegExp('##PATH##', 'g'), context.asAbsolutePath('.'));
+        this.mainTemplate = content.replace(new RegExp('##PATH##', 'g'), extensionPath);
     }
 
-    provideTextDocumentContent(uri: vscode.Uri, token: vscode.CancellationToken): vscode.ProviderResult<string> {
+    private getHtmlContent(uri: vscode.Uri): string {
         var symbolInfo = this.objectLibraries.getBasicLibrary(uri.fsPath, false);
-        var symbolInfoJson = JSON.stringify(symbolInfo);
+        var symbolInfoJson = JSON.stringify(symbolInfo);        
         //build page
         var html = this.mainTemplate.replace('##DATA##', symbolInfoJson);
         return html;
     }
 
-    async appFileObjCommand(objType : string, objId : number, commandName : string) {
+    open() {
+        this.panel = vscode.window.createWebviewPanel("al-code-outline.appViewer", this.title, vscode.ViewColumn.Active, {
+            enableScripts : true
+        });
+
+        this.updateWebView();
+
+        this.panel.onDidDispose(() => this.dispose(), null, this.disposables);
+        this.panel.onDidChangeViewState(e => {
+            if ((this.panel) && (this.panel.visible)) {
+                this.updateWebView();
+            }
+        }, null, this.disposables);
+        this.panel.webview.onDidReceiveMessage(m => {
+            if (m)
+                this.appFileObjCommand(m.objtype, m.objid, m.cmdname);
+        }, null, this.disposables);
+
+    }
+
+    private updateWebView() {
+        if (this.panel)
+            this.panel.webview.html = this.htmlContent;
+    }    
+
+    protected async appFileObjCommand(objType : string, objId : number, commandName : string) {
         if (commandName === 'definition')
             this.objectLibraries.goToDefinition(objType, objId);
         else {