+++ src/extension.ts
@@ -5,8 +5,8 @@ import * as vscode from 'vscode';
 import { ALOutlineProvider } from './aloutline';
 import { ObjectBuildersCollection } from './objectbuilders/objectBuildersCollection';
 import { ALObjectLibrariesCollection } from './objectlibrary/alObjectLibrariesCollection';
-import { ALAppFileContentProvider } from './alappviewer/alAppFileContentProvider';
 import { ALObjectRunner } from './alObjectRunner';
+import { ALAppFileViewer } from './alappviewer/alAppFileViewer';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -16,28 +16,7 @@ export function activate(context: vscode.ExtensionContext) {
     
     //app preview
     const objectLibraries = new ALObjectLibrariesCollection();
-    const alAppFileContentProvider = new ALAppFileContentProvider(context, objectLibraries, objectBuildersCollection);
-    const alAppRegistration = vscode.workspace.registerTextDocumentContentProvider('alOutline.appViewer', alAppFileContentProvider);
     
-    context.subscriptions.push(
-        vscode.commands.registerCommand(
-            'alOutline.appFileObjCommand', 
-            (objtype : string, objid : number, cmdname : string) => {
-                alAppFileContentProvider.appFileObjCommand(objtype, objid, cmdname);
-            }));
-
-    context.subscriptions.push(
-        vscode.commands.registerCommand(
-            'alOutline.viewALApp', 
-            (fileUri) => {
-                let appUri = fileUri.with( {scheme: 'alOutline.appViewer'});
-                vscode.commands.executeCommand('vscode.previewHtml', appUri, vscode.ViewColumn.Active, 'AL Object Browser');
-                //return vscode.commands.executeCommand('vscode.previewHtml', previewUri, vscode.ViewColumn.Two, 'CSS Property Preview').then((success) => {
-                //}, (reason) => {
-                //		vscode.window.showErrorMessage(reason);
-                //});
-            }));
-
     //al code outline
     const alOutlineProvider = new ALOutlineProvider(context);
 
@@ -98,8 +77,15 @@ export function activate(context: vscode.ExtensionContext) {
                 objectRunner.runInWebClient(offset);
             }));
 
-    alAppFileContentProvider.setOutlineProvider(alOutlineProvider);
-    alAppFileContentProvider.setALObjectRunner(objectRunner);
+    context.subscriptions.push(
+        vscode.commands.registerCommand(
+            'alOutline.viewALApp', 
+            (fileUri) => {
+                let appViewer : ALAppFileViewer = new ALAppFileViewer(context, objectLibraries, objectBuildersCollection, 
+                    alOutlineProvider, objectRunner, fileUri);
+                appViewer.open();    
+            }));
+        
 
     alOutlineProvider.refresh();
 }