+++ src/objectlibrary/alObjectLibrariesCollection.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
+import * as path from 'path';
 import { ALObjectLibrary } from "./alObjectLibrary";
 import { ALBasicLibrary } from "./alBasicLibrary";
 import { ALSymbolInfo } from '../alSymbolInfo';
@@ -55,7 +56,79 @@ export class ALObjectLibrariesCollection {
         return null;
     }
 
-    goToDefinition(objectType : string, objectId : number) {
+    protected async runGoToDefinitionOnAlFile(sourceCode : string, posLine : number, posColumn : number, lastSourceLine : number, lastSourceColumn : number) {
+        if ((!vscode.workspace.workspaceFolders) || (vscode.workspace.workspaceFolders.length == 0))
+            return;
+        let fs = require('fs');
+        let tempFileUri = vscode.Uri.file(path.join(vscode.workspace.workspaceFolders[0].uri.fsPath, ".allangtemp", "tempal.al"));
+        let tempFolderUri = path.join(vscode.workspace.workspaceFolders[0].uri.fsPath, ".allangtemp");
+        if (!fs.existsSync(tempFolderUri))
+            fs.mkdirSync(tempFolderUri);
+
+        //create file
+        let edit = new vscode.WorkspaceEdit();
+        edit.createFile(tempFileUri, {overwrite : true});
+        await vscode.workspace.applyEdit(edit);
+
+        //write content to the file
+        edit = new vscode.WorkspaceEdit();
+        edit.insert(tempFileUri, new vscode.Position(0, 0), sourceCode);
+        await vscode.workspace.applyEdit(edit);
+        
+        //download document symbols
+        let pos = new vscode.Position(posLine, posColumn);
+        let list : any = await vscode.commands.executeCommand('vscode.executeDefinitionProvider', tempFileUri, pos);
+
+        //clear file to remove errors from the workspace
+        edit = new vscode.WorkspaceEdit();
+        edit.delete(tempFileUri, new vscode.Range(0, 0, lastSourceLine, lastSourceColumn));
+        await vscode.workspace.applyEdit(edit);
+
+        edit = new vscode.WorkspaceEdit();
+        edit.deleteFile(tempFileUri, {ignoreIfNotExists : true});
+        await vscode.workspace.applyEdit(edit);
+
+        //go to definition
+        if ((list) && (list.length > 0)) {
+            let location : vscode.Location = list[0];
+
+            vscode.workspace.openTextDocument(location.uri).then(
+                document => { 
+                    vscode.window.showTextDocument(document, {
+                        preview : false
+                    });
+                },
+                err => {
+                    vscode.window.showErrorMessage(err);
+                });
+        }
+
+    }
+
+    protected goToAlObjectDefinition(objectType : string, objectName : string) {
+        objectName = '"' + objectName.replace('"', '""') + '"';       
+        let sourceCode : string = 'codeunit 0 "ALLangServerProxy" \n' +
+            '{\n' +
+            'var\n' +
+            'a : ' + objectType + ' ' +  objectName + ';\n' +
+            '}\n';
+        this.runGoToDefinitionOnAlFile(sourceCode, 3, objectType.length + 6, 5, 0);
+    }
+
+    goToDefinition(objectType : string, objectId : number) {       
+        
+        //find object name
+        let symbolInfo : ALSymbolInfo = this.findALSymbolInfo(objectType, objectId);
+        if (symbolInfo == null)
+            return;
+        //translate object type to variable type
+        objectType = objectType.toLowerCase();
+        if (objectType == "table")
+            objectType = "record";
+        
+        this.goToAlObjectDefinition(objectType, symbolInfo.symbolName);       
+        
+        /*
         var uri = this.findObjectUri(objectType, objectId);
         if (uri) {
             vscode.workspace.openTextDocument(uri).then(
@@ -68,6 +141,8 @@ export class ALObjectLibrariesCollection {
                     vscode.window.showErrorMessage(err);
                 });
         }
+        */
+        
     }
 
 }