+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 0.0.10
+
+- github issue [#8 Go to definition no more working](https://github.com/anzwdev/al-code-outline/issues/8) fixed. Go to definition in app file viewer no longer calls al-preview document content provider directly, as it opens empty document. New solution injects al file with empty codeunit with just a single variable declaration into .allangtemp folder, calls standard "go to definition" on that variable type and then removes the file.
+- npm references updated
+
+Thank you KristofKlein for reporting issue #8
+
 ## 0.0.9
 
 - github issue [#7 AL Outline stops working after Create a Page](https://github.com/anzwdev/al-code-outline/issues/7) fixed